gui/core separation of connected components module

+++ ilastik/core/overlayMgr.py
@@ -240,7 +240,7 @@ class OverlayItem(object):
         data = data - dmin
         dmax = numpy.max(data)
         data = 255*data/dmax
-        data = data.astype(numpy.uint8) # transform to uint8
+        data = data.astype(numpy.uint32) # transform to uint8
         return data
 
     @classmethod
